% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_alters_followers.R
\name{generate_alters_followers}
\alias{generate_alters_followers}
\title{Function to get followers of followers from a previously generated ego's list of followers}
\usage{
generate_alters_followers(ego_network, degree_n, n_alters = NULL,
  n_alters_alters = NULL)
}
\arguments{
\item{ego_network}{name of the ego network df (to minimize the number of API calls, you must pass in the ego network, not simply the ego user's name)}

\item{degree_n}{degree that these followers are connected to the original ego}

\item{n_alters}{number of ego's alters to retrieve followers for (set to NULL to retrieve followers for all alters. Value of NULL is likely to result in extremely long run time)}

\item{n_alters_alters}{max num of each alter's followers to retrieve (set to NULL to retrieve all followers for each alter. Value of NULL is likely to result in extremely long run time)}
}
\description{
Get networks of each alter following the original ego
}
\examples{
generate_alters_followers(APSA_network)
}
