% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ego_follower_edgelist.R,
%   R/generate_ego_friend_edgelist.R
\name{generate_ego_follower_edgelist}
\alias{generate_ego_follower_edgelist}
\alias{generate_ego_follower_edgelist}
\title{Function to get all followers of an ego account and return a dataframe of all directed edges between ego and followers}
\usage{
generate_ego_follower_edgelist(ego_name, n = NULL)

generate_ego_follower_edgelist(ego_name, n = NULL)
}
\arguments{
\item{n}{number of followers to retrieve, default of NULL will retrieve all followers. If a user has fewer than n followers, all followers will be retrieved}

\item{ego}{ego username}

\item{ego}{ego username}

\item{n}{number of friends to retrieve, default of NULL will retrieve all followers. If a user has fewer than n friends, all friends will be retrieved}
}
\description{
Function to get all followers of an ego account and return a dataframe of all directed edges between ego and followers

Function to get all friends of an ego account and return a dataframe of all directed edges between ego and friends
}
\examples{
generate_ego_follower_edgelist("APSAtweets")
generate_ego_friends_edgelist("APSAtweets")
}
