% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ego_friend_edgelist.R
\name{generate_ego_friend_edgelist}
\alias{generate_ego_friend_edgelist}
\title{Function to get all friends of an ego account and return a dataframe of all directed edges between ego and friends}
\usage{
generate_ego_friend_edgelist(ego_name, n = NULL)
}
\arguments{
\item{n}{number of friends to retrieve, default of NULL will retrieve all followers. If a user has fewer than n friends, all friends will be retrieved}

\item{ego}{ego username}
}
\description{
Function to get all friends of an ego account and return a dataframe of all directed edges between ego and friends
}
\examples{
generate_ego_friend_edgelist("APSAtweets")
}
