% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ego_follower_edgelist.R
\name{generate_ego_follower_edgelist}
\alias{generate_ego_follower_edgelist}
\title{Function to get all followers of an ego account and return a dataframe of all directed edges between ego and followers}
\usage{
generate_ego_follower_edgelist(ego_name, n = NULL)
}
\arguments{
\item{n}{number of followers to retrieve, default of NULL will retrieve all followers. If a user has fewer than n followers, all followers will be retrieved}

\item{ego}{ego username}
}
\description{
Function to get all followers of an ego account and return a dataframe of all directed edges between ego and followers
}
\examples{
generate_ego_follower_edgelist("APSAtweets")
}
